+++ build-info-extractor-gradle/src/main/groovy/org/jfrog/gradle/plugin/artifactory/ArtifactoryPlugin.groovy
@@ -4,9 +4,11 @@ import org.gradle.api.Project
 import org.jfrog.gradle.plugin.artifactory.dsl.ArtifactoryPluginConvention
 import org.jfrog.gradle.plugin.artifactory.task.ArtifactoryTask
 import org.jfrog.gradle.plugin.artifactory.task.DeployTask
+import org.jfrog.gradle.plugin.artifactory.task.ExtractModuleTask
 
 import static org.jfrog.gradle.plugin.artifactory.task.ArtifactoryTask.ARTIFACTORY_PUBLISH_TASK_NAME
 import static org.jfrog.gradle.plugin.artifactory.task.ArtifactoryTask.DEPLOY_TASK_NAME
+import static org.jfrog.gradle.plugin.artifactory.task.ArtifactoryTask.EXTRACT_MODULE_TASK_NAME
 
 /**
  * @author Lior Hasson
@@ -28,6 +30,13 @@ class ArtifactoryPlugin extends ArtifactoryPluginBase {
     protected DeployTask createArtifactoryDeployTask(Project project) {
         def result = project.getTasks().create(DEPLOY_TASK_NAME, DeployTask.class)
         result.setDescription('''Deploys artifacts and build-info to Artifactory''')
+        return result
+    }
+
+    @Override
+    protected ExtractModuleTask createExtractModuleTask(Project project) {
+        def result = project.getTasks().create(EXTRACT_MODULE_TASK_NAME, ExtractModuleTask.class)
+        result.setDescription('''Extracts module info to an intermediate file''')
         return result
     }
 }
