+++ build-info-extractor-gradle/src/main/groovy/org/jfrog/gradle/plugin/artifactory/extractor/GradleBuildInfoExtractor.java
@@ -16,76 +16,41 @@
 
 package org.jfrog.gradle.plugin.artifactory.extractor;
 
-import com.google.common.base.Function;
-import com.google.common.base.Predicate;
-import com.google.common.collect.Iterables;
-import com.google.common.collect.Sets;
 import org.apache.commons.lang.StringUtils;
 import org.gradle.api.Project;
-import org.gradle.api.Task;
-import org.gradle.api.artifacts.Configuration;
-import org.gradle.api.artifacts.ModuleVersionIdentifier;
-import org.gradle.api.artifacts.ResolvedArtifact;
-import org.gradle.api.artifacts.ResolvedConfiguration;
-import org.gradle.api.internal.artifacts.configurations.ConfigurationInternal;
-import org.gradle.api.internal.artifacts.configurations.DefaultConfiguration;
+import org.gradle.api.artifacts.ProjectDependency;
 import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
-import org.gradle.api.tasks.TaskState;
-import org.gradle.internal.concurrent.GradleThread;
-import org.gradle.util.SingleMessageLogger;
 import org.jfrog.build.api.Agent;
-import org.jfrog.build.api.Artifact;
 import org.jfrog.build.api.BlackDuckProperties;
 import org.jfrog.build.api.Build;
 import org.jfrog.build.api.BuildAgent;
 import org.jfrog.build.api.BuildType;
-import org.jfrog.build.api.Dependency;
 import org.jfrog.build.api.Governance;
 import org.jfrog.build.api.Issue;
 import org.jfrog.build.api.IssueTracker;
 import org.jfrog.build.api.Issues;
 import org.jfrog.build.api.LicenseControl;
 import org.jfrog.build.api.MatrixParameter;
-import org.jfrog.build.api.Module;
 import org.jfrog.build.api.Vcs;
-import org.jfrog.build.api.builder.ArtifactBuilder;
 import org.jfrog.build.api.builder.BuildInfoBuilder;
-import org.jfrog.build.api.builder.DependencyBuilder;
-import org.jfrog.build.api.builder.ModuleBuilder;
 import org.jfrog.build.api.builder.PromotionStatusBuilder;
 import org.jfrog.build.api.release.Promotion;
-import org.jfrog.build.api.util.FileChecksumCalculator;
 import org.jfrog.build.extractor.BuildInfoExtractor;
 import org.jfrog.build.extractor.BuildInfoExtractorUtils;
+import org.jfrog.build.extractor.ModuleExtractorUtils;
 import org.jfrog.build.extractor.clientConfiguration.ArtifactoryClientConfiguration;
-import org.jfrog.build.extractor.clientConfiguration.IncludeExcludePatterns;
-import org.jfrog.build.extractor.clientConfiguration.PatternMatcher;
-import org.jfrog.build.extractor.clientConfiguration.deploy.DeployDetails;
-import org.jfrog.gradle.plugin.artifactory.ArtifactoryPluginUtil;
 import org.jfrog.gradle.plugin.artifactory.task.ArtifactoryTask;
 
-import javax.annotation.Nullable;
 import java.io.File;
-import java.lang.reflect.Method;
+import java.io.IOException;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
-import java.util.List;
 import java.util.Map;
 import java.util.Properties;
 import java.util.Set;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ExecutionException;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-
-import static com.google.common.collect.Iterables.*;
-import static com.google.common.collect.Lists.newArrayList;
-import static org.jfrog.build.extractor.BuildInfoExtractorUtils.getModuleIdString;
-import static org.jfrog.build.extractor.BuildInfoExtractorUtils.getTypeString;
 
 /**
  * An upload task uploads files to the repositories assigned to it.  The files that get uploaded are the artifacts of
@@ -95,9 +60,8 @@ import static org.jfrog.build.extractor.BuildInfoExtractorUtils.getTypeString;
  */
 public class GradleBuildInfoExtractor implements BuildInfoExtractor<Project> {
     private static final Logger log = Logging.getLogger(GradleBuildInfoExtractor.class);
+    public static final String ALL_MODULES_CONFIGURATION = "allModules";
 
-    private static final String SHA1 = "sha1";
-    private static final String MD5 = "md5";
     private final ArtifactoryClientConfiguration clientConf;
     private final Set<GradleDeployDetails> gradleDeployDetails;
     private int publishForkCount;
@@ -145,20 +109,25 @@ public class GradleBuildInfoExtractor implements BuildInfoExtractor<Project> {
         long durationMillis = buildStartDate != null ? System.currentTimeMillis() - buildStartDate.getTime() : 0;
         bib.durationMillis(durationMillis);
 
-        Set<Project> allProjects = rootProject.getAllprojects();
-        if (publishForkCount <= 1) {
-            allProjects.forEach(p -> addModule(bib, p));
-        } else {
-            try {
-                ExecutorService executor = Executors.newFixedThreadPool(publishForkCount);
-                CompletableFuture<Void> allModules = CompletableFuture.allOf(allProjects.stream()
-                        .map(project -> CompletableFuture.runAsync(() -> addModule(bib, project), executor))
-                        .toArray(CompletableFuture[]::new));
-                allModules.get();
-            } catch (InterruptedException | ExecutionException e) {
-                throw new RuntimeException(e);
+        Set<File> moduleFilesWithModules = rootProject.getConfigurations().getByName(ALL_MODULES_CONFIGURATION).files(dependency -> {
+            if (dependency instanceof ProjectDependency) {
+                Project dependencyProject = ((ProjectDependency) dependency).getDependencyProject();
+                if (dependencyProject.getState().getExecuted()) {
+                    ArtifactoryTask artifactoryTask = ProjectUtils.getBuildInfoTask(dependencyProject);
+                    return artifactoryTask != null && artifactoryTask.hasModules();
+                }
             }
-        }
+
+            return false;
+        });
+        moduleFilesWithModules.forEach(moduleFile -> {
+            try {
+                bib.addModule(ModuleExtractorUtils.readModuleFromFile(moduleFile));
+            } catch (IOException e) {
+                throw new RuntimeException("Cannot load module info from file: " + moduleFile.getAbsolutePath(), e);
+            }
+        });
+
         String parentName = clientConf.info.getParentBuildName();
         String parentNumber = clientConf.info.getParentBuildNumber();
         if (parentName != null && parentNumber != null) {
@@ -270,188 +239,5 @@ public class GradleBuildInfoExtractor implements BuildInfoExtractor<Project> {
             build.setParentBuildId(parentName);
         }
         return build;
-    }
-
-    private Module extractModuleInManagedThread(Project project) {
-        try {
-            GradleThread.setManaged();
-            return extractModule(project);
-        } finally {
-            GradleThread.setUnmanaged();
-        }
-    }
-
-    private void addModule(BuildInfoBuilder bib, Project project) {
-        if (project.getState().getExecuted()) {
-            ArtifactoryTask buildInfoTask = getBuildInfoTask(project);
-            if (buildInfoTask != null && buildInfoTask.hasModules()) {
-                bib.addModule(extractModuleInManagedThread(project));
-            }
-        }
-    }
-
-    private ArtifactoryTask getBuildInfoTask(Project project) {
-        Set<Task> tasks = project.getTasksByName(ArtifactoryTask.ARTIFACTORY_PUBLISH_TASK_NAME, false);
-        if (tasks.isEmpty()) {
-            return null;
-        }
-        ArtifactoryTask artifactoryTask = (ArtifactoryTask)tasks.iterator().next();
-        if (taskDidWork(artifactoryTask)) {
-            return artifactoryTask;
-        }
-        return null;
-    }
-
-    /**
-     * Determines if the task actually did any work.
-     * This methods wraps Gradle's task.getState().getDidWork().
-     * @param task  The ArtifactoryTask
-     * @return      true if the task actually did any work.
-     */
-    private boolean taskDidWork(ArtifactoryTask task) {
-        try {
-            return task.getState().getDidWork();
-        } catch (NoSuchMethodError error) {
-            // Compatibility with older versions of Gradle:
-            try {
-                Method m = task.getClass().getMethod("getState");
-                TaskState state = (TaskState)m.invoke(task);
-                return state.getDidWork();
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            }
-        }
-    }
-
-    public Module extractModule(Project project) {
-        String artifactName = project.getName();
-        ArtifactoryTask task = getBuildInfoTask(project);
-        if (task != null) {
-            artifactName = project.getName();
-        }
-        ModuleBuilder builder = new ModuleBuilder()
-                .id(getModuleIdString(project.getGroup().toString(),
-                        artifactName, project.getVersion().toString()));
-        try {
-            ArtifactoryClientConfiguration.PublisherHandler publisher = ArtifactoryPluginUtil.getPublisherHandler(project);
-            if (publisher != null) {
-                boolean excludeArtifactsFromBuild = publisher.isFilterExcludedArtifactsFromBuild();
-                IncludeExcludePatterns patterns = new IncludeExcludePatterns(
-                        publisher.getIncludePatterns(),
-                        publisher.getExcludePatterns());
-                Iterable<GradleDeployDetails> deployExcludeDetails = null;
-                Iterable<GradleDeployDetails> deployIncludeDetails = null;
-                if (excludeArtifactsFromBuild) {
-                    deployIncludeDetails = Iterables.filter(gradleDeployDetails, new IncludeExcludePredicate(project, patterns, true));
-                    deployExcludeDetails = Iterables.filter(gradleDeployDetails, new IncludeExcludePredicate(project, patterns, false));
-                } else {
-                    deployIncludeDetails = Iterables.filter(gradleDeployDetails, new ProjectPredicate(project));
-                    deployExcludeDetails = new ArrayList<GradleDeployDetails>();
-                }
-                builder.artifacts(calculateArtifacts(deployIncludeDetails))
-                        .excludedArtifacts(calculateArtifacts(deployExcludeDetails))
-                        .dependencies(calculateDependencies(project));
-            } else {
-                log.warn("No publisher config found for project: " + project.getName());
-            }
-        } catch (Exception e) {
-            log.error("Error during extraction: ", e);
-        }
-        return builder.build();
-    }
-
-    private List<Artifact> calculateArtifacts(Iterable<GradleDeployDetails> deployDetails) throws Exception {
-        List<Artifact> artifacts = newArrayList(transform(deployDetails, new Function<GradleDeployDetails, Artifact>() {
-            public Artifact apply(GradleDeployDetails from) {
-                PublishArtifactInfo publishArtifact = from.getPublishArtifact();
-                DeployDetails deployDetails = from.getDeployDetails();
-                String artifactPath = deployDetails.getArtifactPath();
-                int index = artifactPath.lastIndexOf('/');
-                return new ArtifactBuilder(artifactPath.substring(index + 1))
-                        .type(getTypeString(publishArtifact.getType(),
-                                publishArtifact.getClassifier(), publishArtifact.getExtension()))
-                        .md5(deployDetails.getMd5()).sha1(deployDetails.getSha1()).build();
-            }
-        }));
-        return artifacts;
-    }
-
-    private List<Dependency> calculateDependencies(Project project) throws Exception {
-        Set<Configuration> configurationSet = project.getConfigurations();
-        List<Dependency> dependencies = newArrayList();
-        for (Configuration configuration : configurationSet) {
-            DefaultConfiguration defaultConfiguration = (DefaultConfiguration) configuration;
-            if (defaultConfiguration.getResolvedState().compareTo(ConfigurationInternal.InternalState.GRAPH_RESOLVED) < 0) {
-                log.info("Artifacts for configuration '{}' were not all resolved, skipping", configuration.getName());
-                continue;
-            }
-            ResolvedConfiguration resolvedConfiguration = SingleMessageLogger.whileDisabled(configuration::getResolvedConfiguration);
-            Set<ResolvedArtifact> resolvedArtifactSet = resolvedConfiguration.getResolvedArtifacts();
-            for (final ResolvedArtifact artifact : resolvedArtifactSet) {
-                File file = artifact.getFile();
-                if (file != null && file.exists()) {
-                    ModuleVersionIdentifier id = artifact.getModuleVersion().getId();
-                    final String depId = getModuleIdString(id.getGroup(),
-                            id.getName(), id.getVersion());
-                    Predicate<Dependency> idEqualsPredicate = new Predicate<Dependency>() {
-                        public boolean apply(@Nullable Dependency input) {
-                            return input.getId().equals(depId);
-                        }
-                    };
-                    // if it's already in the dependencies list just add the current scope
-                    if (any(dependencies, idEqualsPredicate)) {
-                        Dependency existingDependency = find(dependencies, idEqualsPredicate);
-                        Set<String> existingScopes = existingDependency.getScopes();
-                        existingScopes.add(configuration.getName());
-                        existingDependency.setScopes(existingScopes);
-                    } else {
-                        DependencyBuilder dependencyBuilder = new DependencyBuilder()
-                                .type(getTypeString(artifact.getType(),
-                                        artifact.getClassifier(), artifact.getExtension()))
-                                .id(depId)
-                                .scopes(Sets.newHashSet(configuration.getName()));
-                        if (file.isFile()) {
-                            // In recent gradle builds (3.4+) subproject dependencies are represented by a dir not jar.
-                            Map<String, String> checksums = FileChecksumCalculator.calculateChecksums(file, MD5, SHA1);
-                            dependencyBuilder.md5(checksums.get(MD5)).sha1(checksums.get(SHA1));
-                        }
-                        dependencies.add(dependencyBuilder.build());
-                    }
-                }
-            }
-        }
-        return dependencies;
-    }
-
-    private class ProjectPredicate implements Predicate<GradleDeployDetails> {
-        private final Project project;
-
-        private ProjectPredicate(Project project) {
-            this.project = project;
-        }
-
-        public boolean apply(@Nullable GradleDeployDetails input) {
-            return input.getProject().equals(project);
-        }
-    }
-
-    private class IncludeExcludePredicate implements Predicate<GradleDeployDetails> {
-        private Project project;
-        private IncludeExcludePatterns patterns;
-        private boolean include;
-
-        public IncludeExcludePredicate(Project project, IncludeExcludePatterns patterns, boolean isInclude) {
-            this.project = project;
-            this.patterns = patterns;
-            include = isInclude;
-        }
-
-        public boolean apply(@Nullable GradleDeployDetails input) {
-            if (include) {
-                return input.getProject().equals(project) && !PatternMatcher.pathConflicts(input.getDeployDetails().getArtifactPath(), patterns);
-            } else {
-                return input.getProject().equals(project) && PatternMatcher.pathConflicts(input.getDeployDetails().getArtifactPath(), patterns);
-            }
-        }
     }
 }
